% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best.H.R
\name{best.H}
\alias{best.H}
\title{Pick the best covariate for split.
This is for the split one node into two nodes in the DTRtree considering only current step.}
\usage{
best.H(H, A, mus.hat, minsplit = 20, w_vec, weight_combine)
}
\arguments{
\item{H}{A matrix of covariates before assigning final treatment, excluding previous treatment variables.}

\item{A}{A vector of observed treatment options.}

\item{mus.hat}{Estimated conditional mean outcome.}

\item{minsplit}{Minimal node size.}

\item{w_vec}{A matrix of weight vectors used to aggregate gain values across
outcomes.}

\item{weight_combine}{Rule used to aggregate gain values across \code{w_vec}.}
}
\description{
Pick the best covariate for split.
This is for the split one node into two nodes in the DTRtree considering only current step.
}
