% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOGTree-package.R, R/M.propen.R
\name{M.propen}
\alias{M.propen}
\title{Additional documentation for \code{M.propen()}}
\usage{
M.propen(A, Xs)
}
\arguments{
\item{A}{Treatment vector.}

\item{Xs}{Covariate matrix.}
}
\value{
\code{M.propen()} returns a propensity score matrix with one column per
treatment level and one row per subject.
}
\description{
Additional documentation for \code{M.propen()}

Estimate propensity score by a multinomial model.
}
\examples{
set.seed(2)
X <- data.frame(X1 = rnorm(30), X2 = rnorm(30))
A <- sample(1:3, 30, replace = TRUE)
M.propen(A, X)
}
