% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOGTree-package.R
\name{mogtree-plot-helpers}
\alias{mogtree-plot-helpers}
\alias{bin2ten}
\alias{ten2bin}
\alias{vis_tree}
\alias{my_vis}
\alias{my_vis_pair}
\alias{my_vis_elipse}
\alias{my_vis_mean}
\alias{my_vis_mean_pair}
\alias{my_vis_all_pair}
\title{Internal plotting and tree-display helpers}
\usage{
ten2bin(x, max_h)

bin2ten(vec)

vis_tree(max_h, treeout, number)

my_vis(mu1.hat, mu2.hat, chosen, k, label, lwd = 3)

my_vis_pair(mu1.hat, mu2.hat, chosen1, chosen2, k, label, lwd = 3)

my_vis_elipse(mu1.hat, mu2.hat, chosen, k, ref_k, label, lwd = 3, textsize)

my_vis_mean(mu1.hat, mu2.hat, chosen, k, label, lwd = 3)

my_vis_mean_pair(mu1.hat, mu2.hat, chosen1, chosen2, i, j, label, k,
  point_size = 3, my_magick = c("bricks", "horizontalsaw", "verticalsaw",
  "hs_horizontal", "hs_vertical"))

my_vis_all_pair(mu1.hat, mu2.hat, leaf_pred, label, k, alpha = 1,
  my_magick = c("bricks", "horizontalsaw", "verticalsaw", "hs_horizontal",
  "hs_vertical"))
}
\description{
These functions are retained to support the analysis scripts in the
repository. They are not part of the stable user-facing API.
}
\details{
Several plotting helpers expect the calling script to define
objects such as \code{Custom.color} and \code{textsize}.
}
\keyword{internal}
