% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Split.X.R
\name{Split.X}
\alias{Split.X}
\title{Split a parent node into two child nodes by a covariate.
Split the parent node in order to maximize the outcome of interest by signing different treatment in child nodes. Calculate new outcome means for each child node.}
\usage{
Split.X(X, A, mus.hat, w_vec, minsplit = 20, weight_combine)
}
\arguments{
\item{X}{A vector of a covariate for each patients.}

\item{A}{A vector of observed treatment indicators.}

\item{mus.hat}{Estimated conditional mean outcome.}

\item{w_vec}{A matrix of weight vectors used to aggregate gain values across
outcomes.}

\item{minsplit}{Minimal node size.}

\item{weight_combine}{Rule used to aggregate gain values across \code{w_vec}.}
}
\description{
Split a parent node into two child nodes by a covariate.
Split the parent node in order to maximize the outcome of interest by signing different treatment in child nodes. Calculate new outcome means for each child node.
}
